# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mutiAI Core API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runtime/bindings": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "List Runtime Bindings",
        "operationId": "list_runtime_bindings_api_v1_runtime_bindings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuntimeBindingResponse"
                  },
                  "type": "array",
                  "title": "Response List Runtime Bindings Api V1 Runtime Bindings Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runtime/bindings/{binding_key}": {
      "put": {
        "tags": [
          "runtime"
        ],
        "summary": "Put Runtime Binding",
        "operationId": "put_runtime_binding_api_v1_runtime_bindings__binding_key__put",
        "parameters": [
          {
            "name": "binding_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Binding Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeBindingUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeBindingResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorEnvelope"
      },
...
      "RuntimeBindingResponse": {
        "properties": {
          "runtime_binding_id": {
            "type": "string",
            "title": "Runtime Binding Id"
          },
          "binding_key": {
            "type": "string",
            "title": "Binding Key"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "security_mode": {
            "$ref": "#/components/schemas/RuntimeSecurityMode"
          },
          "capability_profile": {
            "$ref": "#/components/schemas/RuntimeCapabilityProfileResponse"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "runtime_binding_id",
          "binding_key",
          "provider",
          "model",
          "reasoning_effort",
          "security_mode",
          "capability_profile",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "RuntimeBindingResponse"
      },
      "RuntimeBindingUpsertRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "security_mode": {
            "$ref": "#/components/schemas/RuntimeSecurityMode"
          },
          "capability_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeCapabilityProfileSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "provider",
          "security_mode"
        ],
        "title": "RuntimeBindingUpsertRequest"
      },
      "RuntimeCapabilityProfileResponse": {
        "properties": {
          "capability_profile_id": {
            "type": "string",
            "title": "Capability Profile Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "profile": {
            "$ref": "#/components/schemas/RuntimeCapabilityProfileSpec"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "trusted": {
            "type": "boolean",
            "title": "Trusted"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "capability_profile_id",
          "revision",
          "profile",
          "source",
          "trusted",
          "observed_at",
          "created_at"
        ],
        "title": "RuntimeCapabilityProfileResponse"
      },
      "RuntimeCapabilityProfileSpec": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "os_family": {
            "type": "string",
            "enum": [
              "linux",
              "windows",
              "macos",
              "unknown"
            ],
            "title": "Os Family",
            "default": "unknown"
          },
          "os_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Os Version"
          },
          "architecture": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "headless": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headless"
          },
          "cpu_capacity_class": {
            "type": "string",
            "enum": [
              "light",
              "standard",
              "heavy",
              "unknown"
            ],
            "title": "Cpu Capacity Class",
            "default": "unknown"
          },
          "memory_mb": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Mb"
          },
          "gpu_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Available"
          },
          "gpu_kind": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Kind"
          },
          "gpu_memory_mb": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Memory Mb"
          },
          "installed_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Installed Tools",
            "default": []
          },
          "tool_inventory_complete": {
            "type": "boolean",
            "title": "Tool Inventory Complete",
            "default": false
          },
          "network_access": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Access"
          },
          "external_services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "External Services",
            "default": []
          },
          "external_service_inventory_complete": {
            "type": "boolean",
            "title": "External Service Inventory Complete",
            "default": false
          },
          "attached_hardware": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attached Hardware",
            "default": []
          },
          "hardware_inventory_complete": {
            "type": "boolean",
            "title": "Hardware Inventory Complete",
            "default": false
          },
          "proprietary_software": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Proprietary Software",
            "default": []
          },
          "proprietary_software_inventory_complete": {
            "type": "boolean",
            "title": "Proprietary Software Inventory Complete",
            "default": false
          },
          "supported_input_media_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Input Media Types",
            "default": []
          },
          "supported_output_media_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Output Media Types",
            "default": []
          },
          "media_inventory_complete": {
            "type": "boolean",
            "title": "Media Inventory Complete",
            "default": false
          },
          "max_duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Duration Seconds"
          },
          "max_input_size_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RuntimeCapabilityProfileV1",
        "description": "Versioned declaration of an effective Runtime environment."
      },
...
      "RuntimeSecurityMode": {
        "type": "string",
        "enum": [
          "demo_full_access",
          "workspace_restricted"
        ],
        "title": "RuntimeSecurityMode",
        "description": "Named product security profiles compiled into Runtime-specific policy."
      },
...
    }
  }
}